+++ src/StorageBase/Bridge/Support/AbstractSingletonStorageFacade.php
@@ -7,6 +7,8 @@ namespace Heptacom\HeptaConnect\Storage\Base\Bridge\Support;
 use Heptacom\HeptaConnect\Storage\Base\Bridge\Contract\StorageFacadeInterface;
 use Heptacom\HeptaConnect\Storage\Base\Bridge\Contract\StorageFacadeServiceExceptionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Bridge\Exception\StorageFacadeServiceException;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\FileReference\FileReferenceGetRequestActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\FileReference\FileReferencePersistRequestActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityMapActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityOverviewActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityPersistActionInterface;
@@ -52,6 +54,10 @@ use Heptacom\HeptaConnect\Storage\Base\Contract\StorageKeyGeneratorContract;
 
 abstract class AbstractSingletonStorageFacade implements StorageFacadeInterface
 {
+    private ?FileReferenceGetRequestActionInterface $fileReferenceGetRequestAction = null;
+
+    private ?FileReferencePersistRequestActionInterface $fileReferencePersistRequestAction = null;
+
     private ?IdentityErrorCreateActionInterface $identityErrorCreateAction = null;
 
     private ?IdentityMapActionInterface $identityMapAction = null;
@@ -133,6 +139,28 @@ abstract class AbstractSingletonStorageFacade implements StorageFacadeInterface
     private ?WebHttpHandlerConfigurationFindActionInterface $webHttpHandlerConfigurationFindAction = null;
 
     private ?WebHttpHandlerConfigurationSetActionInterface $webHttpHandlerConfigurationSetAction = null;
+
+    public function getFileReferenceGetRequestAction(): FileReferenceGetRequestActionInterface
+    {
+        try {
+            return $this->fileReferenceGetRequestAction ??= $this->createFileReferenceGetRequestAction();
+        } catch (StorageFacadeServiceExceptionInterface $throwable) {
+            throw $throwable;
+        } catch (\Throwable $throwable) {
+            throw new StorageFacadeServiceException(FileReferenceGetRequestActionInterface::class, $throwable);
+        }
+    }
+
+    public function getFileReferencePersistRequestAction(): FileReferencePersistRequestActionInterface
+    {
+        try {
+            return $this->fileReferencePersistRequestAction ??= $this->createFileReferencePersistRequestAction();
+        } catch (StorageFacadeServiceExceptionInterface $throwable) {
+            throw $throwable;
+        } catch (\Throwable $throwable) {
+            throw new StorageFacadeServiceException(FileReferencePersistRequestActionInterface::class, $throwable);
+        }
+    }
 
     public function getIdentityErrorCreateAction(): IdentityErrorCreateActionInterface
     {
@@ -571,6 +599,16 @@ abstract class AbstractSingletonStorageFacade implements StorageFacadeInterface
             throw new StorageFacadeServiceException(WebHttpHandlerConfigurationSetActionInterface::class, $throwable);
         }
     }
+
+    /**
+     * @throws \Throwable
+     */
+    abstract protected function createFileReferenceGetRequestAction(): FileReferenceGetRequestActionInterface;
+
+    /**
+     * @throws \Throwable
+     */
+    abstract protected function createFileReferencePersistRequestAction(): FileReferencePersistRequestActionInterface;
 
     /**
      * @throws \Throwable