+++ src/DatasetBase/CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `\Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait::isAttached` to check for a specific instance of an object in the attachment list
 - Add `\Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait::detach` to remove a specific instance from the attachment list
 - Add `\Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface` to match the trait `\Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait` and add it to `\Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract`
+- Add class `\Heptacom\HeptaConnect\Dataset\Base\File\FileReferenceContract` as a base class for various file reference implementations
+- Add class `\Heptacom\HeptaConnect\Dataset\Base\File\FileReferenceCollection` as a collection for `\Heptacom\HeptaConnect\Dataset\Base\File\FileReferenceContract`
 
 ### Changed